Regorganized auto upgrade code
